Refactor test print functionality into independent methods
